backend/requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
filelock==3.19.1
fsspec==2025.9.0
geoclip==1.2.0
geographiclib==2.1
geopy==2.4.1
huggingface-hub==0.35.3
idna==3.11
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
packaging==25.0
pandas==2.3.3
pillow==11.3.0
fastapi==0.115.2
pydantic==2.10.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
reverse_geocode==1.6.6
safetensors==0.6.2
scipy==1.15.3
six==1.17.0
sympy==1.13.1
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn[standard]==0.32.0
